requirements.txt
Django==1.11
djangorestframework==3.6.2
Pillow==7.1.1
pytz==2019.3
